update is_atom function to follow PEP8
"The preferred way of wrapping long lines is by using Python's implied line continuation inside parentheses, brackets and braces. Long lines can be broken over multiple lines by wrapping expressions in parentheses. These should be used in preference to using a backslash for line continuation."
and
"The preferred place to break around a binary operator is after the operator, not before it."

[Source](http://legacy.python.org/dev/peps/pep-0008/#maximum-line-length)